+++ src/extension.ts
@@ -1,31 +1,31 @@
 'use strict';
 import * as vscode from 'vscode';
 
-function getInsertCode(mode:number, variableName:string, prefix:string, suffix:string, attr:string, color?:string) {
+function getInsertCode(printfunction:string, usefstring:boolean, mode:number, variableName:string, prefix:string, suffix:string, attr:string, color?:string) {
     let codeToInsert = '';
-    if (mode === 0) {
-        if(color) {
-            // codeToInsert = `print(colored("${prefix}${variableName}${attr}${suffix}", "${color}"), ${variableName}${attr})`;
-            codeToInsert = `print(colored(f"${prefix}${variableName}${attr}${suffix}{${variableName}${attr}}", "${color}"))`;
-        } else {
-            // codeToInsert = `print("${prefix}${variableName}${attr}${suffix}", ${variableName}${attr})`;
-            codeToInsert = `print(f"${prefix}${variableName}${attr}${suffix}{${variableName}${attr}}")`;
+    if (usefstring){
+        if (mode === 0) { // design for ctrl shift o
+            if(color) {
+                codeToInsert = `${printfunction}(colored(f"${prefix}${variableName}${attr}${suffix}{${variableName}${attr}}", "${color}"))`;
+            } else {
+                codeToInsert = `${printfunction}(f"${prefix}${variableName}${attr}${suffix}{${variableName}${attr}}")`;
+            }
+        } else if (mode === 1) { // design for ctrl shift t
+            if(color) {
+                codeToInsert = `${printfunction}(colored(f"${prefix}${attr}(${variableName})${suffix}{${attr}(${variableName})}", "${color}"))`;
+            } else {
+                codeToInsert = `${printfunction}(f"${prefix}${attr}(${variableName})${suffix}{${attr}(${variableName})}")`;
+            }
         }
-    } else if (mode === 1) {
-        if(color) {
-            // codeToInsert = `print(colored("${prefix}${attr}(${variableName})${suffix}", "${color}"), ${attr}(${variableName}))`;
-            codeToInsert = `print(colored(f"${prefix}${attr}(${variableName})${suffix}{${attr}(${variableName})}", "${color}"))`;
-        } else {
-            // codeToInsert = `print("${prefix}${attr}(${variableName})${suffix}", ${attr}(${variableName}))`;
-            codeToInsert = `print(f"${prefix}${attr}(${variableName})${suffix}{${attr}(${variableName})}")`;
+        else if (mode === 2) {
+            if(color) {
+                codeToInsert = `${printfunction}(colored("${variableName}".center(${prefix}, "${attr}"), "${color}"))`;
+            } else {
+                codeToInsert = `${printfunction}("${variableName}".center(${prefix}, "${attr}"))`;
+            }
         }
-    }
-    else if (mode === 2) {
-        if(color) {
-            codeToInsert = `print(colored("${variableName}".center(${prefix}, "${attr}"), "${color}"))`;
-        } else {
-            codeToInsert = `print("${variableName}".center(${prefix}, "${attr}"))`;
-        }
+    }else{
+        codeToInsert = `${printfunction}(${variableName})`;
     }
     return codeToInsert;
 }
@@ -40,27 +40,13 @@ let insertText = (text: string, moveCursor?: boolean) => {
 
     editor.edit((editBuilder) => editBuilder.insert(selection.start, text)).then(() => {
       if (moveCursor) {
-        // After editBuilder.insert(), the cursor have moved to the end of line
-
-        // You have 2 methods to move the cursor to the destination position
-        
-        // Method 1: use built-in command 'cursorMove'
+        // After editBuilder.insert(), the cursor have moved to the end of line, then move the cursor to the destination position
         vscode.commands.executeCommand('cursorMove', { to: 'left' });
-
-        // Method 2 have 2 sub-methods, first one need get newselection, second one need calculate text length
-        // Method 2-1:
-        // let newselection = editor.selection;
-        // var destPosition = new vscode.Position(newselection.start.line, newselection.start.character - 1);
-        // Method 2-2:
-        // var destPosition = new vscode.Position(selection.start.line, selection.start.character + text.length - 1);  // selection.start.character is indent
-
-        // var newSelection = new vscode.Selection(destPosition, destPosition);
-        // editor.selection = newSelection;
       }
     });
 };
 
-async function handleInsertion(prefix:string, suffix:string, attr:string, mode:number, color?:string) {
+async function handleInsertion(printfunction:string, usefstring:boolean, prefix:string, suffix:string, attr:string, mode:number, color?:string) {
     let editor = vscode.window.activeTextEditor;
     if (!editor) {
         vscode.window.showErrorMessage('Can\'t insert print() because no python file is opened or cursor is not focused');
@@ -81,12 +67,12 @@ async function handleInsertion(prefix:string, suffix:string, attr:string, mode:n
             console.log("regexList[0]:" + regexList[0]);
             console.log("regexList[1]:" + regexList[1]);
             if (regexList[0].includes(',') && /^[A-Za-z0-9_,\.\s]+$/.test(regexList[0])) {  // the regex is to avoid regard sth like 'm[1,2]' as unpacked variable
-                // there are multiple unpakced variables1,2
+                // there are multiple unpakced variables
                 let v_list = regexList[0].split(',');
                 for (var variableName of v_list){
                     await vscode.commands.executeCommand('editor.action.insertLineAfter')
                     .then(() => {
-                        const codeToInsert = getInsertCode(mode, variableName.trim(), prefix, suffix, attr, color);
+                        const codeToInsert = getInsertCode(printfunction, usefstring, mode, variableName.trim(), prefix, suffix, attr, color);
                         insertText(codeToInsert);
                     });
                 }
@@ -95,7 +81,7 @@ async function handleInsertion(prefix:string, suffix:string, attr:string, mode:n
                 let variableName = regexList[0].trim();
                 vscode.commands.executeCommand('editor.action.insertLineAfter')
                 .then(() => {
-                    const codeToInsert = getInsertCode(mode, variableName, prefix, suffix, attr, color);
+                    const codeToInsert = getInsertCode(printfunction, usefstring, mode, variableName, prefix, suffix, attr, color);
                     insertText(codeToInsert);
                 });
             }
@@ -121,7 +107,7 @@ async function handleInsertion(prefix:string, suffix:string, attr:string, mode:n
             for (var variableName of v_list){
                 await vscode.commands.executeCommand('editor.action.insertLineAfter')
                 .then(() => {
-                    const codeToInsert = getInsertCode(mode, variableName.trim(), prefix, suffix, attr, color);
+                    const codeToInsert = getInsertCode(printfunction, usefstring, mode, variableName.trim(), prefix, suffix, attr, color);
                     insertText(codeToInsert);
                 });
             }
@@ -129,13 +115,13 @@ async function handleInsertion(prefix:string, suffix:string, attr:string, mode:n
             const linecontent = editor.document.lineAt(selection.start.line).text.trim();
             // If selected variable is new defined for the first time, insert code at current line, or insert at next line
             if(selected_text === linecontent) {
-                const codeToInsert = getInsertCode(mode, selected_text, prefix, suffix, attr, color);
+                const codeToInsert = getInsertCode(printfunction, usefstring, mode, selected_text, prefix, suffix, attr, color);
                 // to achieve insert in current line, just replace the selected text with codeToInsert
                 editor.edit((editBuilder) => editBuilder.replace(selection, codeToInsert));
             } else {
                 vscode.commands.executeCommand('editor.action.insertLineAfter')
                 .then(() => {
-                    const codeToInsert = getInsertCode(mode, selected_text, prefix, suffix, attr, color);
+                    const codeToInsert = getInsertCode(printfunction, usefstring, mode, selected_text, prefix, suffix, attr, color);
                     insertText(codeToInsert);
                 });
             }
@@ -171,106 +157,95 @@ async function handleCommentOut(mode:string) {
                 end = totalCount;
                 break;
         }
-        for (let i = start; i < end+1; i++) {
-            let line = editor.document.lineAt(i);
-            // comment the line
-            if (line.text.trim().startsWith("print(") || line.text.trim().startsWith("prints(")) {
-                // why use await, see:https://github.com/Microsoft/vscode/issues/9874#issuecomment-235769379
-                await editor.edit(editBuilder => {
-                    // get the indent position
-                    const indent = line.firstNonWhitespaceCharacterIndex;
-                    const p = new vscode.Position(i, indent);
-                    editBuilder.insert(p, '# ');
-                });
-            }
-        }
     } else {
         start = editor.selection.start.line;
         end = editor.selection.end.line;
-        // toggle
-        for (let i = start; i < end+1; i++) {
-            let line = editor.document.lineAt(i);
-            // comment the line
-            if (line.text.trim().startsWith("print(") || line.text.trim().startsWith("prints(")) {
-                // why use await, see:https://github.com/Microsoft/vscode/issues/9874#issuecomment-235769379
-                await editor.edit(editBuilder => {
-                    // get the indent position
-                    const indent = line.firstNonWhitespaceCharacterIndex;
-                    const p = new vscode.Position(i, indent);
-                    editBuilder.insert(p, '# ');
-                });
-            } else if (line.text.trim().startsWith("# print(") || line.text.trim().startsWith("# prints(")) {
-                // uncomment out the line
-                await editor.edit(editBuilder => {
-                    const indent = line.firstNonWhitespaceCharacterIndex;
-                    let toDeleteRange = new vscode.Range(line.range.start.translate(0,indent), line.range.start.translate(0, indent+2));
-                    editBuilder.delete(toDeleteRange);
-                }
-                );
+    }
+    // toggle
+    for (let i = start; i < end+1; i++) {
+        let line = editor.document.lineAt(i);
+        const lineText = line.text.trim();
+        // comment the line
+        if (lineText.startsWith("print(") || lineText.startsWith("prints(")) {
+            await editor.edit(editBuilder => {
+                // get the indent position
+                const indent = line.firstNonWhitespaceCharacterIndex;
+                const p = new vscode.Position(i, indent);
+                editBuilder.insert(p, '# ');
+            });
+        } else if (lineText.startsWith("# print(") || lineText.startsWith("# prints(")) {
+            // uncomment the line
+            await editor.edit(editBuilder => {
+                const indent = line.firstNonWhitespaceCharacterIndex;
+                let toDeleteRange = new vscode.Range(line.range.start.translate(0,indent), line.range.start.translate(0, indent+2));
+                editBuilder.delete(toDeleteRange);
             }
+            );
         }
     }
-    
-
 }
 
 export function activate(context: vscode.ExtensionContext) {
     console.log('Quick Python Print is now active!');
 
-    let prefix = vscode.workspace.getConfiguration().get('1.1.prefix');
-    let suffix = vscode.workspace.getConfiguration().get('1.2.suffix');
-    let attr1 = vscode.workspace.getConfiguration().get('2.attribute1');
-    let attr2 = vscode.workspace.getConfiguration().get('3.attribute2');
-    let builtinfunc = vscode.workspace.getConfiguration().get('4.function');
-    let colortext = vscode.workspace.getConfiguration().get('5.1.enable colored output text');
-    let color1 = vscode.workspace.getConfiguration().get('5.2.color for ctrl shift l');
-    let color2 = vscode.workspace.getConfiguration().get('5.3.color for ctrl shift o');
-    let color3 = vscode.workspace.getConfiguration().get('5.4.color for ctrl shift t');
-    let delimierSymbol = vscode.workspace.getConfiguration().get('6.1.delimiter symbol for ctrl shift ;');
-    let delimierLength = vscode.workspace.getConfiguration().get('6.2.delimiter length for ctrl shift ;');
-    let delimierColor = vscode.workspace.getConfiguration().get('6.3.delimiter color for ctrl shift ;');
+    let printfunction:string = vscode.workspace.getConfiguration().get('0.print function');
+    let usefstring:boolean = vscode.workspace.getConfiguration().get('1.enable formatted string');
+    let prefix:string = vscode.workspace.getConfiguration().get('2.prefix');
+    let suffix:string = vscode.workspace.getConfiguration().get('3.suffix');
+    let attr1:string = vscode.workspace.getConfiguration().get('4.1.attribute1');
+    let attr2:string = vscode.workspace.getConfiguration().get('4.2.attribute2');
+    let builtinfunc:string = vscode.workspace.getConfiguration().get('5.function');
+    let colortext:boolean = vscode.workspace.getConfiguration().get('6.1.enable colored output text');
+    let color1:string = vscode.workspace.getConfiguration().get('6.2.color for ctrl shift l');
+    let color2:string = vscode.workspace.getConfiguration().get('6.3.color for ctrl shift o');
+    let color3:string = vscode.workspace.getConfiguration().get('6.4.color for ctrl shift t');
+    let delimierSymbol:string = vscode.workspace.getConfiguration().get('7.1.delimiter symbol for ctrl shift ;');
+    let delimierLength:string = vscode.workspace.getConfiguration().get('7.2.delimiter length for ctrl shift ;');
+    let delimierColor:string = vscode.workspace.getConfiguration().get('7.3.delimiter color for ctrl shift ;');
 
     // monitor the configuration changes and update them
     context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(() => {
-        prefix = vscode.workspace.getConfiguration().get('1.1.prefix');
-        suffix = vscode.workspace.getConfiguration().get('1.2.suffix');
-        attr1 = vscode.workspace.getConfiguration().get('2.attribute1');
-        attr2 = vscode.workspace.getConfiguration().get('3.attribute2');
-        builtinfunc = vscode.workspace.getConfiguration().get('4.function');
-        colortext = vscode.workspace.getConfiguration().get('5.1.enable colored output text');
-        color1 = vscode.workspace.getConfiguration().get('5.2.color for ctrl shift l');
-        color2 = vscode.workspace.getConfiguration().get('5.3.color for ctrl shift o');
-        color3 = vscode.workspace.getConfiguration().get('5.4.color for ctrl shift t');
-        delimierSymbol = vscode.workspace.getConfiguration().get('6.1.delimiter symbol for ctrl shift ;');
-        delimierLength = vscode.workspace.getConfiguration().get('6.2.delimiter length for ctrl shift ;');
-        delimierColor = vscode.workspace.getConfiguration().get('6.3.delimiter color for ctrl shift ;');
+        printfunction = vscode.workspace.getConfiguration().get('0.print function');
+        usefstring = vscode.workspace.getConfiguration().get('1.enable formatted string');
+        prefix = vscode.workspace.getConfiguration().get('2.prefix');
+        suffix = vscode.workspace.getConfiguration().get('3.suffix');
+        attr1 = vscode.workspace.getConfiguration().get('4.1.attribute1');
+        attr2 = vscode.workspace.getConfiguration().get('4.2.attribute2');
+        builtinfunc = vscode.workspace.getConfiguration().get('5.function');
+        colortext = vscode.workspace.getConfiguration().get('6.1.enable colored output text');
+        color1 = vscode.workspace.getConfiguration().get('6.2.color for ctrl shift l');
+        color2 = vscode.workspace.getConfiguration().get('6.3.color for ctrl shift o');
+        color3 = vscode.workspace.getConfiguration().get('6.4.color for ctrl shift t');
+        delimierSymbol = vscode.workspace.getConfiguration().get('7.1.delimiter symbol for ctrl shift ;');
+        delimierLength = vscode.workspace.getConfiguration().get('7.2.delimiter length for ctrl shift ;');
+        delimierColor = vscode.workspace.getConfiguration().get('7.3.delimiter color for ctrl shift ;');
     }));
 
     let disposable;
 
     disposable = vscode.commands.registerCommand('extension.python-print', () => {
-        if (Boolean(colortext)) {
-            handleInsertion(String(prefix), String(suffix), String(attr1), 0, String(color1));
+        if (colortext) {
+            handleInsertion(printfunction, usefstring, prefix, suffix, attr1, 0, color1);
         } else {
-            handleInsertion(String(prefix), String(suffix), String(attr1), 0);
+            handleInsertion(printfunction, usefstring, prefix, suffix, attr1, 0);
         }
     });
     context.subscriptions.push(disposable);
 
     disposable = vscode.commands.registerCommand('extension.python-print-tensor-shape', () => {
-        if (Boolean(colortext)) {
-            handleInsertion(String(prefix), String(suffix), String(attr2), 0, String(color2));
+        if (colortext) {
+            handleInsertion(printfunction, usefstring, prefix, suffix, attr2, 0, color2);
         } else {
-            handleInsertion(String(prefix), String(suffix), String(attr2), 0);
+            handleInsertion(printfunction, usefstring, prefix, suffix, attr2, 0);
         }
     });
     context.subscriptions.push(disposable);
 
     disposable = vscode.commands.registerCommand('extension.python-print-function', () => {
-        if (Boolean(colortext)) {
-            handleInsertion(String(prefix), String(suffix), String(builtinfunc), 1, String(color3));
+        if (colortext) {
+            handleInsertion(printfunction, usefstring, prefix, suffix, builtinfunc, 1, color3);
         } else {
-            handleInsertion(String(prefix), String(suffix), String(builtinfunc), 1);
+            handleInsertion(printfunction, usefstring, prefix, suffix, builtinfunc, 1);
         }
     });
     context.subscriptions.push(disposable);
@@ -300,29 +275,20 @@ export function activate(context: vscode.ExtensionContext) {
         let selection = editor.selection;
         if (selection.isEmpty) {
             start = 0;
-            end = editor.document.lineCount-1;  // the lineCount start from 1, so minus 1
+            end = editor.document.lineCount-1;
         } else {
             start = editor.selection.start.line;
             end = editor.selection.end.line;
         }
-        // array to store the lines to be deleted
-        let linesToDelete:number[] = [];
-        for (let i = start; i < end+1; i++) {  // the for loop will not reach end, so plus 1
-            const line = editor.document.lineAt(i);
-            const lineText = line.text.trim();
-            console.log("lineText: " + lineText);
-            if (lineText.startsWith("print(") || lineText.startsWith("# print(") || lineText.startsWith("prints(") || lineText.startsWith("# prints(")) {
-                linesToDelete.push(line.lineNumber);
+        // adapted from https://github.com/wakamex/Quick-Python-Print/blob/3f1bf41358cb9e4e0805b70f9feb2f143f2a4e88/src/extension.ts#L44 
+        await editor.edit(editBuilder => {
+            for (let i = end; i >= start; i--) {
+                if (/^(# )?(print|prints|pprint|logger\.debug|logger\.info)\(/.test(editor.document.lineAt(i).text.trim())) {
+                    editBuilder.delete(editor.document.lineAt(i).rangeIncludingLineBreak);
+                }
             }
-        }
-        console.log("linesToDelete", linesToDelete);
-        // delete all the lines indexed by the array
-        for (let i = 0; i < linesToDelete.length; i++) {
-            const line = editor.document.lineAt(linesToDelete[i]-i);  // Every time a row is deleted, the index is decremented by one
-            await editor.edit(editBuilder => {
-                editBuilder.delete(line.rangeIncludingLineBreak);
-            });
-        }
+        });
+
     });
     context.subscriptions.push(disposable);
 
@@ -335,14 +301,12 @@ export function activate(context: vscode.ExtensionContext) {
         let selection = editor.selection;
         let codeToInsert = "";
         const v:string = "";
-        if (Boolean(colortext)) {
-            codeToInsert = getInsertCode(2, v, String(delimierLength), v, String(delimierSymbol), String(delimierColor));
+        if (colortext) {
+            codeToInsert = getInsertCode(printfunction, usefstring, 2, v, delimierLength, v, delimierSymbol, delimierColor);
         } else {
-            codeToInsert = getInsertCode(2, v, String(delimierLength), v, String(delimierSymbol));
+            codeToInsert = getInsertCode(printfunction, usefstring, 2, v, delimierLength, v, delimierSymbol);
         }
         editor.edit((editBuilder) => editBuilder.insert(selection.start, codeToInsert));
     });
     context.subscriptions.push(disposable);
 }
-
-export function deactivate() {}